Fix floating point precision failures caused by FMA on ppc64le
Following tests fail on ppc64le due to FMA:
xy
    - TestAreaGetCentroid
    - TestCentroid
    - TestSignedArea
    - TestLineGetCentroidLines
    - TestLineGetCentroidPolygons
    - ExampleLinearRingsCentroid

xy/lineintersector
    - TestRobustLineIntersectionLines

xyz
    - TestVectorDot
    - TestVectorLength
    - TestDistanceLineToLine
    - TestDistancePointToLine

See https://go.dev/ref/spec#Floating_point_operators for more details.

+++ xyz/vector.go
@@ -14,7 +14,7 @@ func VectorDot(v1Start, v1End, v2Start, v2End geom.Coord) float64 {
 	v2Startv2Endx := v2End[0] - v2Start[0]
 	v2Startv2Endy := v2End[1] - v2Start[1]
 	v2Startv2Endz := v2End[2] - v2Start[2]
-	return v1Startv2Endx*v2Startv2Endx + v1Startv2Endy*v2Startv2Endy + v1Startv2Endz*v2Startv2Endz
+	return float64(v1Startv2Endx*v2Startv2Endx) + float64(v1Startv2Endy*v2Startv2Endy) + float64(v1Startv2Endz*v2Startv2Endz) // nolint:unconvert
 }
 
 // VectorNormalize creates a coordinate that is the normalized vector from 0,0,0 to vector
@@ -25,5 +25,5 @@ func VectorNormalize(vector geom.Coord) geom.Coord {
 
 // VectorLength calculates the length of the vector from 0,0,0 to vector
 func VectorLength(vector geom.Coord) float64 {
-	return math.Sqrt(vector[0]*vector[0] + vector[1]*vector[1] + vector[2]*vector[2])
+	return math.Sqrt(float64(vector[0]*vector[0]) + float64(vector[1]*vector[1]) + float64(vector[2]*vector[2])) // nolint:unconvert
 }
